% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{addTimeline}
\alias{addTimeline}
\title{Title}
\usage{
addTimeline(map = NULL, data = leaflet::getMapData(map),
  layerId = NULL, group = NULL, timelineOptions = timelineOptions(),
  sliderOptions = sliderOptions(), onchange = NULL)
}
\arguments{
\item{data}{}
}
\description{
Title
}
