% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{leaftimeDependency}
\alias{leaftimeDependency}
\title{'Leaflet.timeline' Dependencies}
\usage{
leaftimeDependency()
}
\value{
\code{htmltools::htmlDependency}
}
\description{
'Leaflet.timeline' Dependencies
}
