% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{sliderOptions}
\alias{sliderOptions}
\title{Title}
\usage{
sliderOptions(start = NULL, end = NULL, position = NULL,
  formatOutput = NULL, enablePlayback = NULL,
  enableKeyboardControls = NULL, steps = NULL, duration = NULL,
  waitToUpdateMap = NULL)
}
\arguments{
\item{waitToUpdateMap}{}
}
\description{
Title
}
