% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{styleOptions}
\alias{styleOptions}
\title{Title}
\usage{
styleOptions(radius = 3, color = NULL, stroke = TRUE, fill = TRUE,
  fillColor = NULL, fillOpacity = NULL)
}
\arguments{
\item{fillOpacity}{}
}
\description{
Title
}
