% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{timelineOptions}
\alias{timelineOptions}
\title{Title}
\usage{
timelineOptions(getInterval = NULL,
  pointToLayer = htmlwidgets::JS("\\nfunction(data, latlng) {\\n  return L.circleMarker(latlng, this.styleOptions);\\n}\\n"),
  styleOptions = styleOptions(), drawOnSetTime = NULL)
}
\arguments{
\item{drawOnSetTime}{}
}
\description{
Title
}
