% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{timelineOptions}
\alias{timelineOptions}
\title{Timeline Options Helper}
\usage{
timelineOptions(
  getInterval = NULL,
  pointToLayer = pointToLayerFun(),
  styleOptions = leaftime::styleOptions(),
  drawOnSetTime = NULL
)
}
\arguments{
\item{getInterval}{\code{htmlwidgets::JS} function that returns an object with
\code{start} and \code{end} properties to specify the start and end
of the timeline range.  See
\href{https://github.com/skeate/Leaflet.timeline#getinterval-function----optional}{getInterval}.}

\item{pointToLayer}{\code{htmlwidgets::JS} function that determines what is drawn
on the map.  By default, a circle marker will be drawn.  See
\href{https://leafletjs.com/reference-1.5.0.html#geojson-pointtolayer}{pointToLayer}.}

\item{styleOptions}{\code{list} from \code{\link{styleOptions}}.}

\item{drawOnSetTime}{\code{logical} to draw when time is set.  Default is \code{TRUE}.  See
\href{https://github.com/skeate/Leaflet.timeline#drawonsettime-boolean----optional-default-true}{drawOnSetTime}.}
}
\value{
\code{list} with options to customize the timeline
}
\description{
Timeline Options Helper
}
\seealso{
\code{\link{addTimeline}}
}
